Added folder exists checks. Added assembly name parameter

+++ ContentReferenceUi/build/post-build.ps1
@@ -1,6 +1,7 @@
 ﻿param(
 	[Parameter(Mandatory=$true)] $projectDir,
-	[Parameter(Mandatory=$true)] $assemblyPath
+	[Parameter(Mandatory=$true)] $assemblyPath,
+	[Parameter(Mandatory=$true)] $assemblyName
 )
 
 function GetModuleVersion($assemblyPath)
@@ -18,8 +19,14 @@ function CreateCorrectlyNamedEmptyExport($projectDir, $moduleVersion)
 	$templateExportPath = $projectDir + "build\EmptyExport.zip"
 	$targetFolder = ("{0}content\App_Data\CMSModules\KenticoCommunity.ContentReferenceUi\Install\" -f $projectDir)
 	$targetExportPath = ("{0}KenticoCommunity.ContentReferenceUi_{1}.zip" -f $targetFolder, $moduleVersion)
-	Write-Host ("      Deleting contents of: {0}" -f $targetFolder)
-	Remove-Item ("{0}*.*" -f $targetFolder)
+	if (Test-Path -Path $targetFolder) {
+		Write-Host ("      Deleting contents of: {0}" -f $targetFolder)
+		Remove-Item ("{0}*.*" -f $targetFolder)
+	}
+    else
+    {
+        New-Item $targetFolder -ItemType Directory
+    }
 	Write-Host ("      Export template: {0}" -f $templateExportPath)
 	Write-Host ("      Target export path: {0}" -f $targetExportPath)
 
@@ -29,6 +36,7 @@ function CreateCorrectlyNamedEmptyExport($projectDir, $moduleVersion)
 try {
 	Write-Host "START: post-build.ps1"
 	Write-Host ("    Project directory: {0}" -f $projectDir)
+	Write-Host ("    Assembly name: {0}" -f $assemblyName)
 	Write-Host ("    Assembly path: {0}" -f $assemblyPath)
 	$moduleVersion = GetModuleVersion $assemblyPath
 	Write-Host ("    Module version: {0}" -f $moduleVersion)
